refactor(desktop): extract markdown emoji popover
Move the inline custom emoji popover renderer into the markdown submodule without changing its implementation.
